refactor : 커스텀 hook
